[Bronze V] Title: 이상한 기호, Time: 212 ms, Memory: 17692 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -6,10 +6,10 @@ public class Main {
 
     Scanner sc = new Scanner(System.in);
 
-    int x = sc.nextInt();
-    int y = sc.nextInt();
-    
-    int result = (x + y) * (x - y);
+    long x = sc.nextInt();
+    long y = sc.nextInt();
+
+    Long result = (x + y) * (x - y);
 
     System.out.println(result);
   }
